+++ aulitech_site/my-app/src/components/Dashboard/UserDevices.jsx
@@ -11,7 +11,6 @@ export default function UserDevices({ devices }) {
     }
   };
 
-  
   const noDevices = (
     <>
       <div>
@@ -74,7 +73,6 @@ export default function UserDevices({ devices }) {
                   </h3>
                 </div>
                 <p className="mt-1 truncate text-lg text-gray-500">
-                  {/* {device.jsondata.operation_mode.value} */}
                 </p>
               </div>
             </div>
@@ -108,5 +106,5 @@ export default function UserDevices({ devices }) {
         </button>
       </div>
     </div>
-  );  
+  );
 }
