+++ aulitech_site/my-app/src/components/GoogleAuth/SignOutAccount.jsx
@@ -3,27 +3,25 @@ import { auth } from "../../firebase";
 import { signOut } from "firebase/auth";
 import { useNavigate } from "react-router-dom";
 
-
 const SignOutAccount = () => {
   const navigate = useNavigate();
-  
   const userSignOut = () => {
     signOut(auth).then(() => {
       console.log('signed out successfully');
       navigate('/')
     })
-    .catch((error) => {
-      console.log("userSignOut error:", error)
-    })
+      .catch((error) => {
+        console.log("userSignOut error:", error)
+      })
   }
 
   return (
     <>
-      <button 
+      <button
         onClick={userSignOut}
         type="button"
         className="decision-button inline-flex rounded-full items-center bg-blue-500 px-2.5 py-1 text-lg font-semibold text-white disabled:bg-gray-200 disabled:cursor-not-allowed hover:opacity-70"
-      >   
+      >
         Sign Out
       </button>
     </>
