+++ aulitech_site/my-app/src/components/NavBar/PracticeMode/PracticeModeToggle.css
@@ -3,33 +3,30 @@
     justify-content: space-between;
     align-items: center;
     padding-right: .5rem;
-    height: 50px; 
+    height: 50px;
 }
 
 .active-device {
-    background-color: #333; /* Active device background */
+    background-color: #333;
 }
 
 .device-name {
-    color: #ddd; /* Text color */
+    color: #ddd;
 }
 
-/* The switch - the box around the slider */
 .switch {
     position: relative;
     display: inline-block;
-    width: 60px; /* Reduced width */
-    height: 20px; /* Reduced height */
+    width: 60px;
+    height: 20px;
 }
 
-/* Hide default HTML checkbox */
 .switch input {
     opacity: 0;
     width: 0;
     height: 0;
 }
 
-/* The slider */
 .slider {
     position: absolute;
     cursor: pointer;
@@ -37,38 +34,38 @@
     left: 0;
     right: 0;
     bottom: 0;
-    background-color: #555; /* Off state */
+    background-color: #555;
     transition: .4s;
 }
 
 .slider:before {
     position: absolute;
     content: "";
-    height: 16px; /* Reduced height */
-    width: 16px; /* Reduced width */
+    height: 16px;
+    width: 16px;
     left: 2px;
     bottom: 2px;
-    top: 2px; 
+    top: 2px;
     background-color: white;
     transition: .4s;
 }
 
-input:checked + .slider {
-    background-color: #a28519; /* On state color */
+input:checked+.slider {
+    background-color: #a28519;
 }
 
-input:focus + .slider {
+input:focus+.slider {
     box-shadow: 0 0 1px #666667;
 }
 
-input:checked + .slider:before {
-    transform: translateX(42px); /* Adjusted for smaller slider */
+input:checked+.slider:before {
+    transform: translateX(42px);
 }
 
-/* Rounded sliders */
 .slider.round {
     border-radius: 34px;
 }
+
 .box {
     display: flex;
     align-items: center;
@@ -80,12 +77,12 @@ input:checked + .slider:before {
 }
 
 .active-device .slider {
-    background-color: #2196F3; /* Active device slider color */
+    background-color: #2196F3;
 }
 
 .practice-mode-text {
     color: rgb(255, 255, 255);
-    font-size: 0.75rem; /* Reduced font size */
+    font-size: 0.75rem;
     margin-left: 10px;
     display: block;
 }
