+++ aulitech_site/my-app/src/components/NavBar/RegisterDevices/RegisterInterface.jsx
@@ -1,18 +1,18 @@
 import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from 'react-router-dom';
 import { collection, doc, getDoc, setDoc, getDocs, setDocs, Firestore, FieldValue, arrayUnion, updateDoc } from "firebase/firestore";
-import { db } from "../../firebase";
-import * as clickerDefault from './cato_schemas/clicker.json';
-import * as mouseDefault from './cato_schemas/mouse.json';
-import * as gestureDefault from './cato_schemas/gesture.json';
-import * as tvRemoteDefault from './cato_schemas/tv_remote.json';
-import * as clickerBindings from '../NavBar/cato_schemas/bindings/clicker_bindings.json';
-import * as tvRemoteBindings from '../NavBar/cato_schemas/bindings/tv_remote_bindings.json';
-import * as gestureMouseBindings from '../NavBar/cato_schemas/bindings/gesture_mouse_bindings.json';
+import { db } from "../../../firebase";
+import * as clickerDefault from '../cato_schemas/clicker.json';
+import * as mouseDefault from '../cato_schemas/mouse.json';
+import * as gestureDefault from '../cato_schemas/gesture.json';
+import * as tvRemoteDefault from '../cato_schemas/tv_remote.json';
+import * as clickerBindings from '../cato_schemas/bindings/clicker_bindings.json';
+import * as tvRemoteBindings from '../cato_schemas/bindings/tv_remote_bindings.json';
+import * as gestureMouseBindings from '../cato_schemas/bindings/gesture_mouse_bindings.json';
 
-import * as practiceDefault from './cato_schemas/practice.json';
-import * as connectionSpecificDefault from './cato_schemas/connection_specific.json';
-import * as operationDefault from './cato_schemas/operation.json';
+import * as practiceDefault from '../cato_schemas/practice.json';
+import * as connectionSpecificDefault from '../cato_schemas/connection_specific.json';
+import * as operationDefault from '../cato_schemas/operation.json';
 
 const deepCopy = (obj) => {
   return JSON.parse(JSON.stringify(obj));
@@ -31,9 +31,7 @@ const RegisterInterface = ({ user, devices }) => {
   console.log('thisDevice', thisDevice);
 
   const [interfaceName, setInterfaceName] = useState("");
-  //const [bluetoothId, setBluetoothId] = useState("");
   const [isInterfaceFocused, setIsInterfaceFocused] = useState(false);
-  //const [isBluetoothFocused, setIsBluetoothFocused] = useState(false);
   const [isOpModeFocused, setIsOpModeFocused] = useState(false);
   const [operationMode, setOperationMode] = useState("");
   const [userCatosList, setUserCatosList] = useState([]); // this is the list of all the nicknames of userCatos
@@ -51,8 +49,6 @@ const RegisterInterface = ({ user, devices }) => {
   if (!userId) {
     console.log("No user ID available");
   }
-
-
 
   const getDocID = async (userId, index) => {
     try {
@@ -104,14 +100,13 @@ const RegisterInterface = ({ user, devices }) => {
     }
   }, [userDeviceData]);
 
-
   const focusStyle = {
     borderColor: '#AA9358',
     boxShadow: '0 0 0 2px #AA9358',
   };
 
   const handleSave = async () => {
-    console.log("Save button clicked"); //debug
+    console.log("Save button clicked"); 
 
     if (interfaceName.trim() === "" || operationMode.trim() === "") {
       alert("Please fill out all fields before saving.");
@@ -123,7 +118,7 @@ const RegisterInterface = ({ user, devices }) => {
     try {
 
       const getConnections = async () => {
-        //parse thru and check if int name already exists TODO 
+        // parse thru and check if int name already exists TODO 
         // get the list of the current connection names for thisDevice
         let currentConnections = [];
         for (let i = 0; i < thisDevice.data.connections.length; i++) {
@@ -156,7 +151,6 @@ const RegisterInterface = ({ user, devices }) => {
         let gestureMouseData = {}
         let practiceData = {}
 
-
         //iterate through connectionSpecificDefault and add the fields to combinedData
         for (const [key, value] of Object.entries(connectionSpecificDefault)) {
           connectionData[key] = value;
@@ -169,7 +163,6 @@ const RegisterInterface = ({ user, devices }) => {
         let clickerOperation = deepCopy(operationDefault);
         clickerOperation.operation_mode.value = 'clicker';
         clickerData = {
-          // ...connectionData,
           ...clickerOperation,
           ...clickerDefault,
           ...clickerBindings,
@@ -192,7 +185,6 @@ const RegisterInterface = ({ user, devices }) => {
         let tvRemoteOperation = deepCopy(operationDefault);
         tvRemoteOperation.operation_mode.value = 'tv_remote';
         tvRemoteData = {
-          // ...connectionData,
           ...tvRemoteOperation,
           ...tvRemoteDefault,
           ...gestureDefault,
@@ -246,8 +238,6 @@ const RegisterInterface = ({ user, devices }) => {
         ]);
 
         console.log("Connection registered successfully");
-
-
       };
       await getConnections();
       navigate(`/devices/${deviceName}`);
@@ -300,7 +290,6 @@ const RegisterInterface = ({ user, devices }) => {
                 style={isInterfaceFocused ? focusStyle : null}
               />
             </div>
-
 
             <div className="mt-5">
               <label htmlFor="op-mode" className="block text-lg text-gray-900">
