--- aulitech_site/my-app/src/components/NavBar/UserSettings.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const UserSettings = () => {
-  return (
-    <div>
-      <h1>User Settings</h1>
-      <p>This is the User Settings page where users can manage their account settings.</p>
-      {/* Add more content here as needed */}
-    </div>
-  );
-}
-
-export default UserSettings;
