+++ aulitech_site/my-app/src/components/NavBar/images/Logo.jsx
@@ -3,7 +3,7 @@ const Logo = ({height, marginY, marginX}) => {
     <div className={`logo-container my-${marginY} mx-${marginX}`}>
       <img
         className={`h-${height} w-auto`}
-        src={require("../../images/icononly_transparent_nobuffer.png")}
+        src={require("../../../images/icononly_transparent_nobuffer.png")}
         alt="Auli logo"
       />
     </div>
