--- aulitech_site/my-app/src/components/RecordGests/DeviceAccess.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function DeviceAccess() {
-  
-}
