+++ aulitech_site/my-app/src/components/RecordGests/RecordGestures.jsx
@@ -2,7 +2,6 @@ import React, { useState, useRef } from "react";
 import ProgBar from "./record/ProgBar";
 import RcrdCard from "./record/RcrdCard";
 
-
 const RecordGestures = ({
   classNames,
   gestName,
@@ -21,7 +20,6 @@ const RecordGestures = ({
     } else {
       setStepCount(stepCount + 1);
       setWriteConnect(false);
-      // setCheckConnect(false);
     }
   };
