--- aulitech_site/my-app/src/components/RecordGests/cors.json
@@ -1,8 +1,0 @@
-[
-    {
-      "origin": ["http://localhost:3000"], 
-      "method": ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
-      "maxAgeSeconds": 3600
-    }
-  ]
-
